+++ gps_timemachine/test/test_gps.py
@@ -11,7 +11,7 @@ def test_gps_to_utc_start():
     gps_time = 0.0
     expected_dt = dt.datetime(1981, 1, 6, 0, 0, 0)
 
-    actual_dt = gps_to_utc(d, gps_time)
+    actual_dt = gps_to_utc(d, gps_time, True)
 
     assert_equal(expected_dt, actual_dt)
 
@@ -22,7 +22,7 @@ def test_gps_to_utc_early_years():
     gps_time = 0.0
     expected_dt = dt.datetime(1982, 7, 1, 23, 59, 58)
 
-    actual_dt = gps_to_utc(d, gps_time)
+    actual_dt = gps_to_utc(d, gps_time, True)
 
     assert_equal(expected_dt, actual_dt)
 
@@ -33,7 +33,7 @@ def test_gps_to_utc_middle_years():
     gps_time = 0.0
     expected_dt = dt.datetime(1999, 12, 31, 23, 59, 47)
 
-    actual_dt = gps_to_utc(d, gps_time)
+    actual_dt = gps_to_utc(d, gps_time, True)
 
     assert_equal(expected_dt, actual_dt)
 
@@ -44,7 +44,7 @@ def test_gps_to_utc_later_years():
     gps_time = 0.0
     expected_dt = dt.datetime(2017, 1, 1, 23, 59, 42)
 
-    actual_dt = gps_to_utc(d, gps_time)
+    actual_dt = gps_to_utc(d, gps_time, True)
 
     assert_equal(expected_dt, actual_dt)
 
@@ -55,7 +55,7 @@ def test_gps_to_utc_later_years_with_hms():
     gps_time = 123456.789
     expected_dt = dt.datetime(2017, 1, 2, 12, 34, 38, 789000)
 
-    actual_dt = gps_to_utc(d, gps_time)
+    actual_dt = gps_to_utc(d, gps_time, True)
 
     assert_equal(expected_dt, actual_dt)
 
@@ -66,7 +66,7 @@ def test_gps_to_utc_large_milliseconds_bug():
     gps_time = 123456.9996
     expected_dt = dt.datetime(2017, 1, 2, 12, 34, 39)
 
-    actual_dt = gps_to_utc(d, gps_time)
+    actual_dt = gps_to_utc(d, gps_time, True)
 
     assert_equal(expected_dt, actual_dt)
 
@@ -77,13 +77,13 @@ def test_gps_to_utc_large_milliseconds_rolls_to_next_day():
     gps_time = 235959.9996
     expected_dt = dt.datetime(1981, 1, 7, 0, 0, 0)
 
-    actual_dt = gps_to_utc(d, gps_time)
+    actual_dt = gps_to_utc(d, gps_time, True)
 
     assert_equal(expected_dt, actual_dt)
 
 
 def test_load_leap_seconds():
-    leap_seconds = load_leap_seconds()
+    leap_seconds = load_leap_seconds(True)
     expected = (dt.datetime(1983, 7, 1, 0, 0), 3.0)
     actual = leap_seconds[25]